true x64!!! wow, revolutionary!

+++ X64Emulator/x64emulator.cpp
@@ -20,7 +20,7 @@ private:
     uint64_t r8, r9, r10, r11, r12, r13, r14, r15;
 
     // Memory (simplified as a vector of bytes)
-    std::vector<uint8_t> memory;
+    std::vector<uint64_t> memory;
 
     // Stack (simplified)
     std::vector<uint64_t> stack;
